Multiplication result converted to larger type
CWE-190
CWE-192
CWE-197
CWE-681

Use an explicit cast on one multiplicand so the multiplication is performed in
a wide unsigned type from the start.

Best minimal fix in src/webserver/src/WebServer.cpp
(constructor CAnyImage::CAnyImage(int width, int height)) is to change:

- from: m_alloc_size = width * height * sizeof(uint32) + 0x100;
- to: m_alloc_size = static_cast<unsigned long>(width) *
  static_cast<unsigned long>(height) * sizeof(uint32) + 0x100;

This preserves behavior for valid positive inputs while eliminating
intermediate int overflow in the multiplication.

No new headers or dependencies are required.

+++ src/webserver/src/WebServer.cpp
@@ -926,7 +926,8 @@ CAnyImage::CAnyImage(int width, int height) : m_width(width), m_height(height)
 {
 	m_size = 0;
 	// allocate considering image header
-	m_alloc_size = width * height * sizeof(uint32) + 0x100;
+	m_alloc_size = static_cast<unsigned long>(width) *
+		static_cast<unsigned long>(height) * sizeof(uint32) + 0x100;
 	if ( m_alloc_size ) {
 		m_data = new unsigned char[m_alloc_size];
 	} else {
